refactor: core-s3 서비스 컴포넌트들 default 접근 제어자 사용
- 외부로는 S3Service만 허용하기 위함
