+++ .gitignore
@@ -37,3 +37,4 @@ out/
 .vscode/
 
 application.yml
+zzal/
